petInfoModal and NavBar style
