+++ client/src/components/Navbar/index.jsx
@@ -21,7 +21,7 @@ const AppNavbar = () => {
             <Navbar.Collapse id="responsive-navbar-nav">
               <Nav className='me-auto'>
                 <Nav.Link as={Link} to='/intake'> I Have Pets </Nav.Link>
-                <Nav.Link as={Link} to='/adoption'> Adopt </Nav.Link>
+                <Nav.Link as={Link} to='/adopt'> Adopt </Nav.Link>
                 <Nav.Link as={Link} to='/shopping'>Shopping</Nav.Link>
                 {/* if user is logged in show saved books and logout */}
                 {Auth.loggedIn() ? (
