+++ client/src/utils/seedPetData.js
@@ -1,0 +1,45 @@
+import goodCat from "../assets/imgs/white_kitten.jpg"
+import badCat from "../assets/imgs/bad_kitten.jpg"
+import goodDog from "../assets/imgs/black_labrador.jpg"
+import badDog from "../assets/imgs/white_pom.jpg"
+
+let petInfo = [
+  {
+    title: "Good Cat of the Month",
+    name: "Soju",
+    species: "Cat",
+    color: "White",
+    age: "3 months",
+    gender: "boy",
+    pic: goodCat,
+  },
+  {
+    title: "Bad Cat of the Month",
+    name: "Zoe",
+    species: "Cat",
+    color: "Black",
+    age: "4 months",
+    gender: "girl",
+    pic: badCat,
+  },
+  {
+    title: "Good Dog of the Month",
+    name: "Jack",
+    species: "Dog",
+    color: "Black",
+    age: "7 months",
+    gender: "boy",
+    pic: goodDog,
+  },
+  {
+    title: "Bad Dog of the Month",
+    name: "Joy",
+    species: "Dog",
+    color: "White",
+    age: "10 months",
+    gender: "boy",
+    pic: badDog,
+  },
+]
+
+export default petInfo;
